requirements.txt
anthropic==0.40.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
